Split client and server components.
